requirements.txt
asgiref==3.5.2
boto3==1.26.22
botocore==1.29.22
dj-database-url==1.0.0
Django==4.1.3
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-simple-history==3.2.0
django-storages==1.13.1
django-utils-six==2.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
psycopg2==2.9.5
python-dateutil==2.8.2
rethinkdb==2.4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0